refactor: move code into modules
