+++ index.html
@@ -145,6 +145,6 @@
   <script src="./vendor/mediapipe/drawing_utils.min.js"></script>
 
   <!-- =======================  APP  ======================= -->
-  <script type="module" src="app.js"></script>
+  <script type="module" src="src/main.js"></script>
 </body>
 </html>
